refactor(country_metadata): address PR review comments
- Extract datetime helpers into helper.go (per reviewer request)
- Make GetLocaleByIdentifier private (getLocaleByIdentifier)
- Add named constants for magic strings (numericAuto/Always, relStyleLong/Short/Narrow, dirPast/Future, dateOrderDMY/MDY/YMD)
